Begin extracting routes
